+++ bin/data_channel.dart
@@ -1,3 +1,1 @@
-library data_channel;
-
 export '../lib/data_channel.dart';