+++ lib/data_channel.dart
@@ -1,6 +1,5 @@
 ///
 /// data_channel
 ///
-library data_channel;
 
 export 'package:data_channel/src/main.dart';